+++ src/fabric_cicd/_items/__init__.py
@@ -4,6 +4,7 @@
 from fabric_cicd._items._activator import publish_activators
 from fabric_cicd._items._apacheairflowjob import publish_apacheairflowjobs
 from fabric_cicd._items._copyjob import publish_copyjobs
+from fabric_cicd._items._dataagent import publish_dataagents
 from fabric_cicd._items._dataflowgen2 import publish_dataflows
 from fabric_cicd._items._datapipeline import find_referenced_datapipelines, publish_datapipelines
 from fabric_cicd._items._environment import check_environment_publish_state, publish_environments
@@ -18,6 +19,7 @@ from fabric_cicd._items._manage_dependencies import set_unpublish_order
 from fabric_cicd._items._mirroreddatabase import publish_mirroreddatabase
 from fabric_cicd._items._mounteddatafactory import publish_mounteddatafactories
 from fabric_cicd._items._notebook import publish_notebooks
+from fabric_cicd._items._orgapp import publish_orgapps
 from fabric_cicd._items._report import publish_reports
 from fabric_cicd._items._semanticmodel import publish_semanticmodels
 from fabric_cicd._items._sqldatabase import publish_sqldatabases
@@ -30,6 +32,7 @@ __all__ = [
     "publish_activators",
     "publish_apacheairflowjobs",
     "publish_copyjobs",
+    "publish_dataagents",
     "publish_dataflows",
     "publish_datapipelines",
     "publish_environments",
@@ -43,6 +46,7 @@ __all__ = [
     "publish_mirroreddatabase",
     "publish_mounteddatafactories",
     "publish_notebooks",
+    "publish_orgapps",
     "publish_reports",
     "publish_semanticmodels",
     "publish_sqldatabases",