+++ src/fabric_cicd/changelog.md
@@ -6,6 +6,17 @@ The following contains all major, minor, and patch version release notes.
 -   📝 Documentation Update
 -   ⚡ Internal Optimization
 
+## Version 0.1.30
+
+<span class="md-h2-subheader">Release Date: 2025-10-20</span>
+
+-   ✨ Add support for binding semantic models to on-premise gateways in Fabric workspaces ([#569](https://github.com/microsoft/fabric-cicd/issues/569))
+-   ✨ Add support for publishing and managing Data Agent items ([#556](https://github.com/microsoft/fabric-cicd/issues/556))
+-   ✨ Add OrgApp item type support ([#520](https://github.com/microsoft/fabric-cicd/issues/520))
+-   ⚡ Enhance cross-workspace variable support to allow referencing other attributes ([#583](https://github.com/microsoft/fabric-cicd/issues/583))
+-   🔧 Fix workspace name extraction bug for non-ID attrs using ITEM_ATTR_LOOKUP ([#583](https://github.com/microsoft/fabric-cicd/issues/583))
+-   🔧 Fix capacity requirement check ([#593](https://github.com/microsoft/fabric-cicd/issues/593))
+
 ## Version 0.1.29
 
 <span class="md-h2-subheader">Release Date: 2025-10-01</span>